% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypackage.R
\name{print.indicator_data}
\alias{print.indicator_data}
\title{Print Method for indicator_data Class}
\usage{
\method{print}{indicator_data}(x, ..., width = NULL)
}
\arguments{
\item{x}{An object of class \code{indicator_data}, e.g data_irl}

\item{...}{Further arguments passed to or from other methods}

\item{width}{Adjust the width of the output to fit the console}
}
\description{
Provides a concise overview when an \code{indicator_data} object is printed in the console. Each unique indicator is printed along with its first and last year recorded
}
