% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypackage.R
\name{read_country_indicators}
\alias{read_country_indicators}
\title{Read and process an Indicator CSV File}
\usage{
read_country_indicators(country_code)
}
\arguments{
\item{country_code}{A character string representing the country code (e.g., "irl", "col").
This code is used to construct the filename \code{indicators_COUNTRYCODE.csv}.}
}
\description{
Read and process an Indicator CSV File
}
\details{
This function performs several steps to ensure data quality and usability
\itemize{
\item It uses 'data.table::fread' to read in the large CSV file
\item It automatically skips the first row, which is metadata, important for later functions
\item It uses 'janitor:clean_names' for standardising column names
\item Columns are cleaned using 'stringr::str_squish', ensuring reliable filtering later
\item Numerical columns are coerced to numeric, and character column are coerced to factors.}
The output is given a custom class 'indicator_data' for further S3 methods.
}
