% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypackage.R
\name{summary.indicator_data}
\alias{summary.indicator_data}
\title{Overall summary of the new the new object of 'indicator_class' using skimr}
\usage{
\method{summary}{indicator_data}(object, ...)
}
\arguments{
\item{object}{An "indicator_data" object created by read_indicator_data() function}

\item{...}{Further arguments passed to or from other methods}
}
\description{
Provides a brief statistical summary of the \code{indicator_data} object
}
